+++ src/main/java/net/glowstone/net/codec/play/game/PlayParticleCodec.java
@@ -20,9 +20,9 @@ public final class PlayParticleCodec implements Codec<PlayParticleMessage> {
     public ByteBuf encode(ByteBuf buf, PlayParticleMessage message) throws IOException {
         buf.writeInt(message.getParticle());
         buf.writeBoolean(message.isLongDistance());
-        buf.writeFloat(message.getX());
-        buf.writeFloat(message.getY());
-        buf.writeFloat(message.getZ());
+        buf.writeDouble(message.getX());
+        buf.writeDouble(message.getY());
+        buf.writeDouble(message.getZ());
         buf.writeFloat(message.getOfsX());
         buf.writeFloat(message.getOfsY());
         buf.writeFloat(message.getOfsZ());
